refactor: 팔로잉 패키지 경로 수정
