+++ src/main/java/back/vybz/userservice/follow/application/mongodb/FollowService.java
@@ -1,8 +1,8 @@
-package back.vybz.userservice.user.application.mongodb;
+package back.vybz.userservice.follow.application.mongodb;
 
-import back.vybz.userservice.user.dto.mongodb.request.RequestAddFollowDto;
-import back.vybz.userservice.user.dto.mongodb.request.RequestDeleteFollowDto;
-import back.vybz.userservice.user.dto.mongodb.response.ResponseFollowDto;
+import back.vybz.userservice.follow.dto.mongodb.request.RequestAddFollowDto;
+import back.vybz.userservice.follow.dto.mongodb.request.RequestDeleteFollowDto;
+import back.vybz.userservice.follow.dto.mongodb.response.ResponseFollowDto;
 
 import java.util.List;
