+++ src/main/java/back/vybz/userservice/follow/domain/mongodb/Follower.java
@@ -1,4 +1,4 @@
-package back.vybz.userservice.user.domain.mongodb;
+package back.vybz.userservice.follow.domain.mongodb;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
